// server/prisma/generated/enums.ts
export const GameProviderName = {
  PRAGMATICPLAY: 'PRAGMATICPLAY',
  EVOPLAY: 'EVOPLAY',
  NETENT: 'NETENT',
  PLAYNGO: 'PLAYNGO',
  RELAXGAMING: 'RELAXGAMING',
  HACKSAW: 'HACKSAW',
  BGAMING: 'BGAMING',
  SPRIBE: 'SPRIBE',
  INTERNAL: 'INTERNAL',
  REDTIGER: 'REDTIGER',
  NETGAME: 'NETGAME',
  BIGFISHGAMES: 'BIGFISHGAMES',
  CQNINE: 'CQNINE',
  NOLIMIT: 'NOLIMIT',
  KICKASS: 'KICKASS'
} as const

export type GameProviderName = (typeof GameProviderName)[keyof typeof GameProviderName]


export const ProviderAuthType = {
  API_KEY: 'API_KEY',
  OAUTH2: 'OAUTH2',
  JWT_SIGN: 'JWT_SIGN',
  CUSTOM: 'CUSTOM',
  NONE: 'NONE'
} as const

export type ProviderAuthType = (typeof ProviderAuthType)[keyof typeof ProviderAuthType]